Clean up interface
