+++ test/android/getDependencyConfig.spec.js
@@ -2,31 +2,31 @@ const path = require('path');
 const expect = require('chai').expect;
 const getDependencyConfig = require('../../src/config/android').dependencyConfig;
 const mockFs = require('mock-fs');
-const projects = require('../fixtures/projects');
+const dependencies = require('../fixtures/dependencies');
 
 describe('Config::getDependencyConfig', () => {
 
   before(() => {
-    mockFs(projects);
+    mockFs({ testDir: dependencies });
   });
 
   it('should return an object with android project configuration', () => {
     const userConfig = {};
-    const folder = path.join('testing', 'flat', 'node_modules', 'react-native-vector-icons');
+    const folder = path.join('testDir', 'react-native-vector-icons');
 
     expect(getDependencyConfig(folder, userConfig)).to.be.an('object');
   });
 
   it('should return `null` if android project was not found', () => {
     const userConfig = {};
-    const folder = path.join('testing', 'empty');
+    const folder = path.join('testDir', 'empty');
 
     expect(getDependencyConfig(folder, userConfig)).to.be.null;
   });
 
   it('should return `null` if android project does not contain ReactPackage', () => {
     const userConfig = {};
-    const folder = path.join('testing', 'flat', 'node_modules', 'rnpm-android-no-package');
+    const folder = path.join('testDir', 'rnpm-android-no-package');
 
     expect(getDependencyConfig(folder, userConfig)).to.be.null;
   });
