+++ test/android/readManifest.spec.js
@@ -7,19 +7,19 @@ const projects = require('../fixtures/projects');
 describe('readManifest', () => {
 
   before(() => {
-    mockFs(projects);
+    mockFs({ testDir: projects });
   });
 
   it('should return manifest content if file exists in the folder', () => {
     const manifestPath = path.join(
-      'testing', 'flat', 'android', 'src', 'AndroidManifest.xml'
+      'testDir', 'flat', 'android', 'src', 'AndroidManifest.xml'
     );
 
     expect(readManifest(manifestPath)).to.be.an('object');
   });
 
   it('should throw an error if there is no manifest in the folder', () => {
-    const fakeManifestPath = path.join('testing', 'empty', 'AndroidManifest.xml');
+    const fakeManifestPath = path.join('testDir', 'empty', 'AndroidManifest.xml');
     expect(() => readManifest(fakeManifestPath)).to.throw(Error);
   });
